chore: Extract Single Object encoding related structs from reader/mod.rs
No API breaks!
No functional changes!
